Implement admin user login
When user re-enters app (after closing) they are still logged in (auth.currentUser is set), but could not find a way to get the userType variable to persist. This meant that the options menu on the home screen showed the default options rather than the correct options for the user's account type. Could not work out how to use onSavedInstanceState correctly (answer may be here https://stackoverflow.com/questions/6554317/savedinstancestate-is-always-null). Hacked way around this is to automatically send the user back to the login screen if the user type is unknown, which will then automatically log the user in again and query the users collection for the user's account type again.